+++ fukkatsu/memory/__init__.py
@@ -1,2 +1,0 @@
-from fukkatsu.memory.manage import *
-from fukkatsu.memory.short import *