+++ fukkatsu/observer/__init__.py
@@ -1,1 +1,0 @@
-from fukkatsu.observer.tracker import *